requirements.txt
Flask
gunicorn
gTTS
pdfminer.six
googletrans==4.0.0-rc1
